Split off parts of `__main__.py`
